+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Change log
 
-0.3.1
+0.4.0
 =====
 * [#81](https://github.com/kowainik/relude/issues/81):
   Add `asumMap` to `Foldable` functions.
 * [#89](https://github.com/kowainik/relude/issues/81):
   Add `Relude.Extra.Type` module containing a `typeName` function.
+* [#92](https://github.com/kowainik/relude/issues/92)
+  Add `Relude.Extra.Tuple` module, containing
+  `dupe`, `mapToFst`, `mapToSnd`, and `mapBoth` functions.
+* Remove `openFile` and `hClose`.
+* [#97](https://github.com/kowainik/relude/issues/97):
+  Add `(&&^)` and `(||^)` operators.
 
 0.3.0
 =====